+++ core/model/user.go
@@ -32,12 +32,11 @@ const (
 type Account struct {
 	ID string //this is ID for the account
 
-	Application  Application
-	Organization Organization
+	AppOrg ApplicationOrganization
 
-	Permissions []ApplicationPermission
-	Roles       []ApplicationRole
-	Groups      []ApplicationGroup
+	Permissions []Permission
+	Roles       []AppOrgRole
+	Groups      []AppOrgGroup
 
 	AuthTypes []AccountAuthType
 
@@ -66,9 +65,9 @@ func (a Account) GetAccountAuthType(authTypeID string, identifier string) *Accou
 }
 
 //GetPermissions returns all permissions granted to this account
-func (a Account) GetPermissions() []ApplicationPermission {
+func (a Account) GetPermissions() []Permission {
 	permissionsMap := a.GetPermissionsMap()
-	permissions := make([]ApplicationPermission, len(permissionsMap))
+	permissions := make([]Permission, len(permissionsMap))
 	i := 0
 	for _, permission := range permissionsMap {
 		permissions[i] = permission
@@ -90,8 +89,8 @@ func (a Account) GetPermissionNames() []string {
 }
 
 //GetPermissionsMap returns a map of all permissions granted to this account
-func (a Account) GetPermissionsMap() map[string]ApplicationPermission {
-	permissionsMap := make(map[string]ApplicationPermission, len(a.Permissions))
+func (a Account) GetPermissionsMap() map[string]Permission {
+	permissionsMap := make(map[string]Permission, len(a.Permissions))
 	for _, permission := range a.Permissions {
 		permissionsMap[permission.Name] = permission
 	}